lambda/requirements.txt
Pillow==3.3.1
boto3==1.4.0
botocore==1.4.49
docutils==0.12
futures==3.0.5
jmespath==0.9.0
piexif==1.0.5
python-dateutil==2.5.3
python-magic==0.4.12
s3transfer==0.1.2
six==1.10.0
wsgiref==0.1.2
